processor/requirements.txt
celery[redis]
python-dotenv
boto3
numpy
scipy
librosa
scikit-learn
ffmpeg-python
